A bit of restructuring.
