+++ src/test/java/de/elegal/WordDocumentTests/OldWordDocumentTests.java
@@ -1,0 +1,25 @@
+package de.elegal.WordDocumentTests;
+
+import de.elegal.Documents.WordDocument;
+import org.junit.Assert;
+import org.junit.Test;
+
+import java.util.Set;
+
+import static de.elegal.WordDocumentTests.FinalPaths.*;
+
+public class OldWordDocumentTests {
+
+
+
+    // TODO: Some null check tests
+    // TODO: Write a lot of test cases whether the run splitting is fucked up
+    // TODO: Check for wrong written stuff
+    // TODO: Also with a lot of extra formatting stuff (Endgegner)
+    // TODO: Rightbound
+    // TODO: Check whether __everything__ avaiable is tested
+    // TODO: Add number of replacements in tests
+    // TODO: Multiple replacements of the same tag
+    // TODO: Write one large nice test file
+    // TODO: Next page replacement and tags existing in general
+}
