+++ src/test/java/de/elegal/WordDocumentTests/TODO
@@ -1,0 +1,2 @@
+- Split up Empty Classes
+- setUp and tearDown for any class where it makes sense
